+++ src/app/auth/decorators/permissions.decorator.ts
@@ -1,0 +1,6 @@
+import { SetMetadata } from "@nestjs/common";
+import { UserPermissions } from "src/app/roles/enums/user-permissions.enum";
+
+export const PERMISSION_KEY = 'permissions';
+
+export const RequirePermissions = (...permissions: UserPermissions[]) => SetMetadata(PERMISSION_KEY, permissions);
