--- src/app/auth/models/roles.model.ts
@@ -1,4 +1,0 @@
-export enum Role {
-  CUSTOMER = 'customer',
-  ADMIN = 'admin',
-}
