--- src/app/security/decorators/permissions.decorator.ts
@@ -1,6 +1,0 @@
-import { SetMetadata } from "@nestjs/common";
-import { UserPermissions } from "src/app/roles/enums/user-permissions.enum";
-
-export const PERMISSION_KEY = 'permissions';
-
-export const RequirePermissions = (...permissions: UserPermissions[]) => SetMetadata(PERMISSION_KEY, permissions);
